previous problem fixed and new feature added
previously whenever the front end was started it would present a blank page. this has been fixed through the correction of the the routes router and route have been spelled in the app.js file and a registartion form file has been added however there is a problem suspected to be within the navigation.