working on front end
completted addrecipeform file however unable to test due to a problem with the navgation when logged in as an admin. creating a authcontext file to keep the navbar visible whenever a user goes through hte pages and dynamically change the login and register button to display the user name and logout button when successfully loged in. this file will also authenticate user privilages.